% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{chaussabel_modules}
\alias{chaussabel_modules}
\title{Modules from Chaussabel (2008)}
\format{
A \code{list} of \code{character} of length 260, with one element per
module.
}
\usage{
data(chaussabel_modules)
}
\description{
Repertoire of co-clustering genes. NOTE: Gene Symbols are updated by the updateDataWithLatestHGNCMap.Rmd
when it is run.  Original is kept as orig_chaussabel.
}
\references{
http://www.biir.net/public_wikis/module_annotation/G2_Trial_8_Modules
http://www.nature.com/nri/journal/v14/n4/full/nri3642.html
}
\keyword{datasets}
