% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOUtils.R
\name{.dlSuppFls}
\alias{.dlSuppFls}
\title{Download Supplemental files from GEO}
\usage{
.dlSuppFls(accList, baseDir, study)
}
\arguments{
\item{accList}{accList}

\item{baseDir}{path to base directory (via \code{.getBaseDir()})}

\item{study}{study accession eg \code{SDY269}}
}
\description{
Download, unzip, and return correct paths for GEO supplementary files
}
