% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fix_headers}
\alias{.fix_headers}
\title{fix headers}
\usage{
.fix_headers(ge_list, study)
}
\arguments{
\item{ge_list}{list of matrices for one study}

\item{study}{study accession eg \code{SDY269}}
}
\description{
Custom fixes for individual studies
}
