% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtils.R
\name{.fixHeaders}
\alias{.fixHeaders}
\title{fix headers}
\usage{
.fixHeaders(mxList, study)
}
\arguments{
\item{mxList}{list of matrices for one study}

\item{study}{study accession eg \code{SDY269}}
}
\description{
Custom fixes for individual studies
}
