% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtils.R
\name{.ge_list_to_flat_file}
\alias{.ge_list_to_flat_file}
\title{matrix list to flat file}
\usage{
.ge_list_to_flat_file(ge_list, supp_files_dir, study)
}
\arguments{
\item{ge_list}{list of gene expression tables (matrix, data.frame, etc)}

\item{supp_files_dir}{path to base directory (via \code{.get_supp_files_dir()})}

\item{study}{study accession eg \code{SDY269}}
}
\value{
path to raw, prepped input files
}
\description{
mxList to flat file. Used when samples are in different files on GEO
}
