% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOUtils.R
\name{.get_geo_supp_files}
\alias{.get_geo_supp_files}
\title{Get GEO supp files}
\usage{
.get_geo_supp_files(geo_accession, supp_files_dir)
}
\arguments{
\item{geo_accession}{GSM or GSE accesssion}

\item{supp_files_dir}{path to directory to save supplementary files.
On ImmuneSpace servers, an appropriate directory can be found with
\code{.get_supp_files_dir()}}
}
\description{
Download supplementary files from GEO accession (GSE or GSM)
and return the paths to the files.
}
