% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_supp_files_dir}
\alias{.get_supp_files_dir}
\title{Get supplemental files directory}
\usage{
.get_supp_files_dir(study, gef)
}
\arguments{
\item{study}{study accession eg \code{SDY269}}

\item{gef}{result of ISCon$getDataset("gene_expression_files") for one run.}
}
\description{
Create run-specific base directory for supplementary files.
Important that run specific directory is created to avoid overlap with previous runs.
This done with unique cell_type * arm_accession string (aka cohort_type)
}
