% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCreateMx.R
\name{.getBaseDir}
\alias{.getBaseDir}
\title{Get Base Directory}
\usage{
.getBaseDir(study, gef)
}
\arguments{
\item{study}{study accession eg \code{SDY269}}

\item{gef}{result of ISCon$getDataset("gene_expression_files") for one run.}
}
\description{
Important that run specific directory is created to avoid overlap with previous runs.
This done with unique cell_type * arm_accession string (aka cohort_type)
}
