% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_files.R
\name{.getGEO_custom}
\alias{.getGEO_custom}
\title{getGEO (custom)}
\usage{
.getGEO_custom(geo_accession, supp_files_dir)
}
\arguments{
\item{geo_accession}{GSM or GSE accesssion}

\item{supp_files_dir}{path to directory to save supplementary files.}
}
\description{
Helper to download GEO soft files to supp_files_dir
}
