% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOUtils.R
\name{.getGsmSuppFiles}
\alias{.getGsmSuppFiles}
\title{get GSM Supplementary Files}
\usage{
.getGsmSuppFiles(gsm, baseDir)
}
\arguments{
\item{gsm}{GEO GSM accession}

\item{baseDir}{path to base directory (via \code{.getBaseDir()})}
}
\description{
Download Gsm supplementary files, unzip and return path
}
