% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOUtils.R
\name{.make_id_to_gsm_map}
\alias{.make_id_to_gsm_map}
\title{#' make id to GSM map
#'
#' Create map of GSM accessions to study-given IDs from GEO
#'
#' @param gef result of ISCon$getDataset("gene_expression_files") for one run.
#' @param metaData list of study-specific meta data
#' @param study study accession eg \code{SDY269}
.make_id_to_gsm_map <- function(gef, metaData, study){
mp <- lapply(gef$geo_accession, function(x){
tmp <- GEOquery::getGEO(x)
nm <- tmp@header[\link{metaData$studyIdTerm}][\link{metaData$gsmMapIndex}]
if (study \%in\% names(metaData$smplGsubTerms)) {
nm <- gsub(metaData$smplGsubTerms[\link{study}]$old,
metaData$smplGsubTerms[\link{study}]$new,
nm)
}
return(c(x, nm))
})
mp <- data.frame(do.call(rbind, mp), stringsAsFactors = FALSE)
colnames(mp) <- c("gsm","id")
rownames(mp) <- NULL
return(mp)
}
Make id to GSM map}
\usage{
.make_id_to_gsm_map(
  gsms,
  study_id_term,
  gsm_map_index = 1,
  id_regex_map = NULL,
  supp_files_dir = NULL
)
}
\arguments{
\item{gsms}{Input vector of gsm accession}

\item{study_id_term}{Term from getGEO(gsm) header to use for extracting
sample id. "description" or "title"}

\item{gsm_map_index}{Index returning id in GSM header}

\item{id_regex_map}{list specifying regex terms to map GSE id's using gsub
of the form \code{list(old = "", new = "")}}

\item{supp_files_dir}{path to directory to save supplementary files.}
}
\description{
Create map of GSM accessions to ID's provided in GSE. If needed, appropriate
info for the matrix will be returned by \code{getMetaData}.
}
