% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOUtils.R
\name{.makeIdToGsmMap}
\alias{.makeIdToGsmMap}
\title{make id to GSM map}
\usage{
.makeIdToGsmMap(gef, metaData, study)
}
\arguments{
\item{gef}{result of ISCon$getDataset("gene_expression_files") for one run.}

\item{metaData}{list of study-specific meta data}

\item{study}{study accession eg \code{SDY269}}
}
\description{
Create map of GSM accessions to study-given IDs from GEO
}
