% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtils.R
\name{.map_feature_id_col}
\alias{.map_feature_id_col}
\title{Standardize probe column name}
\usage{
.map_feature_id_col(exprs)
}
\arguments{
\item{exprs}{data.table of expression}
}
\description{
Standardize probe column name
}
