% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmPortUtils.R
\name{.mapAccToBs}
\alias{.mapAccToBs}
\title{Map experiment-sample or geo accessions to biosample accessions}
\usage{
.mapAccToBs(exprs, gef)
}
\arguments{
\item{exprs}{data.table of expression}
}
\description{
Map experiment-sample or geo accessions to biosample accessions
}
