% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtils.R
\name{.mapFeatureIdCol}
\alias{.mapFeatureIdCol}
\title{Standardize probe column name}
\usage{
.mapFeatureIdCol(exprs)
}
\arguments{
\item{exprs}{data.table of expression}
}
\description{
Standardize probe column name
}
