% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtils.R
\name{.mxListToFlatFile}
\alias{.mxListToFlatFile}
\title{matrix list to flat file}
\usage{
.mxListToFlatFile(mxList, baseDir, study)
}
\arguments{
\item{mxList}{matrix list}

\item{baseDir}{path to base directory (via \code{.getBaseDir()})}

\item{study}{study accession eg \code{SDY269}}
}
\value{
path to raw, prepped input files
}
\description{
mxList to flat file. Used when samples are in different files on GEO
}
