% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raw_matrix.R
\name{.prep_illumina_headers}
\alias{.prep_illumina_headers}
\title{Prep Illumina Headers}
\usage{
.prep_illumina_headers(raw_illumina_dt)
}
\arguments{
\item{raw_illumina_dt}{data.table of raw illumina values.}
}
\value{
data.table with cleaned up names ready for \code{read.ilmn()}
}
\description{
For read.ilmn() to work correctly the signal cols need format <smpl>.<exprValTerm>
with the expression-value term being something like "AVG_Signal" or "SAMPLE". Then the
detection p-value cols must have format <smpl>.Detection Pval. This way the
rawElist creates the $E and $other$Detection matrices with same colnames of <smpl>.
Since in this script the read.ilmn(file, exprs = "AVG_Signal", probeid = "ID_REF) is
hardcoded. The vars are substituted here for other viable versions, e.g. SAMPLE and PROBE_ID.
}
