% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOUtils.R
\name{.prepGeoFls}
\alias{.prepGeoFls}
\title{Prep GEO files}
\usage{
.prepGeoFls(study, gef, metaData, inputFiles)
}
\arguments{
\item{study}{study accession eg \code{SDY269}}

\item{gef}{result of ISCon$getDataset("gene_expression_files") for one run.}

\item{metaData}{list of study-specific meta data}

\item{inputFiles}{input file names}
}
\value{
path to raw, prepped input files
}
\description{
Generate flat files that are ready for processing from GEO "raw" data.
Warning! Raw data is highly variable for gse supplementary files
}
