% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmPortUtils.R
\name{.prepImmportFls}
\alias{.prepImmportFls}
\title{Prep ImmPort Files}
\usage{
.prepImmportFls(study, gef, metaData, inputFiles)
}
\arguments{
\item{study}{study accession eg \code{SDY269}}

\item{gef}{result of ISCon$getDataset("gene_expression_files") for one run.}

\item{metaData}{list of study-specific meta data}

\item{inputFiles}{input file names}
}
\value{
path to raw, prepped input files
}
\description{
Prep ImmPort Files
}
