% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Affymetrix.R
\name{.process_affy}
\alias{.process_affy}
\title{Process Affy}
\usage{
.process_affy(input_files)
}
\arguments{
\item{input_files}{input file names}
}
\value{
background-corrected data.table of probe intensities
}
\description{
Process Affy
}
\details{
Requires annotation packages which can be installed from bioconductor.
\code{affy::justRMA()} will use the appropriate package.
Two are customCDF packages loaded from UpdateAnno: \code{huex10stv2cdf}
and \code{hursta2a520709cdf}
}
