% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Illumina.R
\name{.process_illumina}
\alias{.process_illumina}
\title{Process Illumina}
\usage{
.process_illumina(raw_file_path)
}
\arguments{
\item{raw_file_path}{file path to raw illumina file}
}
\description{
For Illumina, correct background using nec() which
calls normexp.fit.control and normexp.signal to use
negative controls as identified by detection p-values
to remove noise.
}
