% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raw_matrix.R
\name{.process_rna_seq}
\alias{.process_rna_seq}
\title{Process RNA-seq}
\usage{
.process_rna_seq(input_files)
}
\arguments{
\item{input_files}{input file names}
}
\description{
Somewhat redundant in case of GEO files as there is no actual processing
of raw counts files ...
}
