% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raw_matrix.R
\name{.process_two_color_array}
\alias{.process_two_color_array}
\title{Process Two Color Array}
\usage{
.process_two_color_array(path)
}
\arguments{
\item{path}{path to}
}
\description{
Two color array processing using limma and assuming genepix files.
bc.method is the background correction method and normexp is used to match
work with Illumina and Affymetrix.
}
