% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Affymetrix.R
\name{.processAffy}
\alias{.processAffy}
\title{Process Affy}
\usage{
.processAffy(inputFiles, gef, metaData)
}
\arguments{
\item{inputFiles}{input file names}

\item{gef}{result of \code{ISCon$getDataset("gene_expression_files")} for one run.}

\item{metaData}{list of study-specific meta data}
}
\value{
background-corrected data.table of probe intensities
}
\description{
Process Affy
}
\details{
Requires annotation packages which can be installed from bioconductor.
\code{affy::justRMA()} will use the appropriate package.
Two are customCDF packages loaded from UpdateAnno: \code{huex10stv2cdf}
and \code{hursta2a520709cdf}
}
