% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Illumina.R
\name{.processIllumina}
\alias{.processIllumina}
\title{Process Illumina}
\usage{
.processIllumina(rawFl)
}
\arguments{
\item{rawFl}{file path to raw illumina file}
}
\description{
For Illumina, correct background using nec() which
calls normexp.fit.control and normexp.signal to use
negative controls as identified by detection p-values
to remove noise.
}
