% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{.processRnaSeq}
\alias{.processRnaSeq}
\title{Process RNA-seq}
\usage{
.processRnaSeq(inputFiles, study)
}
\arguments{
\item{inputFiles}{input file names}

\item{study}{study accession eg \code{SDY269}}
}
\description{
Somewhat redundant in case of GEO files as there is no actual processing
of raw counts files ...
}
