% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoColorArray.R
\name{.processTwoColorArray}
\alias{.processTwoColorArray}
\title{Process Two Color Array}
\usage{
.processTwoColorArray(path)
}
\description{
Two color array processing using limma and assuming genepix files.
bc.method is the background correction method and normexp is used to match
work with Illumina and Affymetrix.
}
