% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_files.R
\name{.select_input_files}
\alias{.select_input_files}
\title{Select input files}
\usage{
.select_input_files(supp_files_dir)
}
\description{
Select the correct input files from supplemental directory and unzip if needed.
}
