% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raw_matrix.R
\name{.subset_raw_illumina_dt}
\alias{.subset_raw_illumina_dt}
\title{Subset Illumina raw data}
\usage{
.subset_raw_illumina_dt(raw_ge_dt)
}
\arguments{
\item{raw_ge_dt}{data.table with raw values from illumina output
for one sample}
}
\description{
Subset Illumina expression matrix to only needed columns
}
