% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Illumina.R
\name{.subsetIlluminaEM}
\alias{.subsetIlluminaEM}
\title{Subset Illumina EM}
\usage{
.subsetIlluminaEM(em)
}
\arguments{
\item{em}{expression}
}
\description{
Subset Illumina expression matrix to only needed columns
}
