% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.write_raw_expression}
\alias{.write_raw_expression}
\title{Write raw expression to a file}
\usage{
.write_raw_expression(ge_tbl, supp_files_dir, study)
}
\arguments{
\item{ge_tbl}{object containing gene expression values (matrix, data.frame, etc)}

\item{supp_files_dir}{path to base directory (via \code{.get_supp_files_dir()})}

\item{study}{study accession eg \code{SDY269}}
}
\value{
path to raw expression file
}
\description{
Writes \code{ge_tbl} to a (tsv-formatted) \code{.txt} file, using controlled
format and file name.
}
