% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtils.R
\name{.write_single_ge_mx}
\alias{.write_single_ge_mx}
\title{write Single Matrix}
\usage{
.write_single_ge_mx(ge_tbl, supp_files_dir, study)
}
\arguments{
\item{ge_tbl}{object containing gene expression values (matrix, data.frame, etc)}

\item{supp_files_dir}{path to base directory (via \code{.get_supp_files_dir()})}

\item{study}{study accession eg \code{SDY269}}
}
\description{
write Single Matrix
}
