% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtils.R
\name{.writeSingleMx}
\alias{.writeSingleMx}
\title{write Single Matrix}
\usage{
.writeSingleMx(em, baseDir, study)
}
\arguments{
\item{em}{matrix of gene expression}

\item{baseDir}{path to base directory (via \code{.getBaseDir()})}

\item{study}{study accession eg \code{SDY269}}
}
\description{
write Single Matrix
}
