% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{emory_blood_transcript_modules}
\alias{emory_blood_transcript_modules}
\title{Blood transcriptome modules}
\format{
A \code{list} of \code{character} of length 346, with one element per
module
}
\usage{
data(emory_blood_transcript_modules)
}
\description{
NOTE: Gene Symbols are updated by the updateDataWithLatestHGNCMap.Rmd when it is run.
Original is kept as orig_emory.
}
\references{
http://www.immuneprofiling.org/meni/meni-paper/btm-landing.gsp
http://www.nature.com/ni/journal/v15/n2/full/ni.2789.html
}
\keyword{datasets}
