% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmx_differential_expression.R
\name{find_de_genes_deseq}
\alias{find_de_genes_deseq}
\title{Compute Differentially Expressed genes for RNAseq using DESeq2}
\usage{
find_de_genes_deseq(eset, p_val_cutoff = 0.02)
}
\arguments{
\item{eset}{expressionset object with raw counts}

\item{p_val_cutoff}{adjusted p-value cutoff for determining differentially
expressed genes.}
}
\description{
Compute Differentially Expressed genes for RNAseq using DESeq2
}
