% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_response_predictor.R
\name{format_predictor_table}
\alias{format_predictor_table}
\title{Format Predictor Table}
\usage{
format_predictor_table(fit)
}
\arguments{
\item{fit}{fit object from \code{glm} or \code{lm}}
}
\description{
Create nicely formatted table of predictors from fit result.
}
