% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_response_predictor.R
\name{get_de_genes}
\alias{get_de_genes}
\title{Get genes differentially expressed over time}
\usage{
get_de_genes(
  con,
  timepoint,
  fc_thresh = 0.58,
  timepoint_unit = "Days",
  cohorts = NULL,
  reload = FALSE
)
}
\arguments{
\item{con}{HMX connection}

\item{timepoint}{Timepoint for finding differentially expressed genes}

\item{fc_thresh}{fold-change threshold for determining whether genes
are differentially expressed}

\item{timepoint_unit}{"Days" or "Hours"}

\item{cohorts}{character vector of cohorts to include}
}
\description{
Get genes differentially expressed over time
}
