% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_response_predictor.R
\name{get_immune_response}
\alias{get_immune_response}
\title{Get immune response.}
\usage{
get_immune_response(
  con,
  assay = "hai",
  participant_ids = NULL,
  dichotomize = FALSE,
  dichotomize_thresh = 4,
  reload = FALSE
)
}
\arguments{
\item{con}{HMX connection}

\item{assay}{"hai", "neut_ab_response", or "elisa"}

\item{participant_ids}{character vector of participant_ids}

\item{dichotomize}{Dichotomize result? If FALSE, max log fold change is
returned. If TRUE, returns TRUE if max log fold change is greater than
\code{dichotomize_value}}

\item{dichotomize_thresh}{Value to use for dichotomizing result if
\code{dichotomize} is TRUE.}
}
\description{
Returns max log fold change of \code{assay} from baseline for each
particiapnt.
}
