% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaData.R
\name{get_meta_data}
\alias{get_meta_data}
\title{Get Meta Data}
\usage{
get_meta_data(study, gef, fas_id)
}
\arguments{
\item{study}{study accession}

\item{gef}{gene expression files}

\item{fas_id}{Feature Annotation Set ID (from FeatureAnnotationSet table)}
}
\description{
Get Meta Data
}
