% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaData.R
\name{getMetaData}
\alias{getMetaData}
\title{Get Meta Data}
\usage{
getMetaData(study, gef, fasId)
}
\arguments{
\item{study}{study accession}

\item{gef}{gene expression files}

\item{fasId}{Feature Annotation Set ID (from FeatureAnnotationSet table)}
}
\description{
Get Meta Data
}
