% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_symbol_mapping.R
\docType{data}
\name{hgncAlias2Symbol}
\alias{hgncAlias2Symbol}
\title{data.table mapping gene aliases to gene symbols}
\format{
data.table with four columns: \describe{
\item{SYMBOL}{"official" gene symbol}
\item{ALIAS}{Aliases for the official gene symbol. Each entry in this
column should be unique}
\item{HGNC}{HGNC ID}
\item{ENTREZ}{Entrez ID}
}
}
\usage{
hgncAlias2Symbol
}
\description{
Gene-alias mapping info from
the HUGO gene nomenclature committee.
}
\details{
Some aliases map to multiple gene symbols. There are a couple of
different cases, which are dealt with differently:
\enumerate{
\item A gene alias maps to multiple gene symbols including itself:
Remove mappings to other symbols.
\item A gene alias maps to multiple gene symbols not including itself:
Remove this alias.
}
}
\references{
\href{https://www.genenames.org/}{HGNC}
}
\keyword{datasets}
