% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_symbol_mapping.R
\docType{data}
\name{hgncAlias2Symbol_version}
\alias{hgncAlias2Symbol_version}
\title{Date hgncAlias2Symbol update}
\format{
An object of class \code{Date} of length 1.
}
\usage{
hgncAlias2Symbol_version
}
\description{
Date hgncAlias2Symbol update
}
\keyword{datasets}
