% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_raw_matrix}
\alias{make_raw_matrix}
\title{Make Raw Matrix}
\usage{
make_raw_matrix(meta_data, gef, study, input_files, analysis_dir)
}
\arguments{
\item{meta_data}{list of study-specific meta data}

\item{gef}{result of ISCon$getDataset("gene_expression_files") for one run.}

\item{study}{study accession eg \code{SDY269}}

\item{input_files}{input file names}
}
\value{
A data.table containing raw counts or background-corrected probe
intensities with a feature_id column and one column per biosample_accession
}
\description{
Make Raw Matrix
}
