% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_annotation.R
\name{makeAnnoDf}
\alias{makeAnnoDf}
\title{Function to map probe ids to gene symbols from an annotation package}
\usage{
makeAnnoDf(annoPkg, outPath = NULL)
}
\arguments{
\item{annoPkg}{name of annotation package to use}

\item{outPath}{file path for tsv to write out for upload to ImmuneSpace}
}
\description{
Function to map probe ids to gene symbols from an annotation package
}
