% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCreateMx.R
\name{makeRawMatrix}
\alias{makeRawMatrix}
\title{Make Raw Matrix}
\usage{
makeRawMatrix(metaData, gef, study, input_files)
}
\arguments{
\item{metaData}{list of study-specific meta data}

\item{gef}{result of ISCon$getDataset("gene_expression_files") for one run.}

\item{study}{study accession eg \code{SDY269}}

\item{input_files}{input file names}
}
\value{
A data.table with a feature_id column and one column per biosample_accession
}
\description{
Make Raw Matrix
}
