% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\docType{data}
\name{meta_data_list}
\alias{meta_data_list}
\title{List of study-specific metadata}
\format{
A list with 4 entries:
\describe{
\item{file_locations}{list of file locations, with a vector of studies
for each of \code{gsm_soft}, \code{gsm_supp_files}, \code{gse_supp_files},
\code{immport}, \code{custom}, \code{not_available}}
\item{id_to_gsm_mapping_info}{list of mapping info for studies which have
data in the GSE and need additional info to map back to the GSM per sample.
Includes: \code{gsm_needs_map}, \code{use_gsm_description},
\code{use_gsm_index_2}, \code{id_regex_map_list}}
\item{illumina_manifest_files}{List with one entry per study which
has an illumina manifest file}
\item{gsm_table_var_name}{list of raw values column name for gsm-based data}
}
}
\usage{
meta_data_list
}
\description{
Summary by study can be returned via \code{\link{get_meta_data}}
}
\keyword{datasets}
