% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{msigdb_immunologic_signatures}
\alias{msigdb_immunologic_signatures}
\title{Immunologic signatures}
\format{
A \code{list} of \code{character} of length 1910, with one element
per module.
}
\usage{
data(msigdb_immunologic_signatures)
}
\description{
Gene sets that represent cell states and perturbations within the immune
system. The signatures were generated by manual curation of published studies
in human and mouse immunology. For each study, pairwise comparisons of
relevant classes were made and genes ranked by mutual information. Gene sets
correspond to top or bottom genes (FDR < 0.25 or maximum of 200 genes) for
each comparison. This resource is generated as part of the Human Immunology
Project Consortium (HIPC; http://www.immuneprofiling.org/). NOTE: Gene Symbols
are updated by the updateDataWithLatestHGNCMap.Rmd when it is run.  Original is kept as
orig_msigdb.
}
\references{
http://www.broadinstitute.org/gsea/msigdb/genesets.jsp?collection=C7
}
\keyword{datasets}
