% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_matrix}
\alias{normalize_matrix}
\title{Normalize matrix}
\usage{
normalize_matrix(exprs_dt, platform, verbose = FALSE)
}
\arguments{
\item{exprs_dt}{data.table with raw, background-corrected expression. One
column per sample, one row per feature.}

\item{platform}{Affymetrix, Illumina, NA, or Stanford Functional Genomics}

\item{verbose}{print verbose logging statements?}
}
\description{
Normalize matrix
}
