% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{normalize_rnaseq}
\alias{normalize_rnaseq}
\title{Normalize raw RNA-seq data}
\usage{
normalize_rnaseq(counts_dt)
}
\arguments{
\item{counts_dt}{data.table with raw counts, one column per sample, one
row per feature}
}
\description{
Normalize raw RNA-seq data
}
