% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCreateMx.R
\name{normalizeMatrix}
\alias{normalizeMatrix}
\title{Normalize Matrix}
\usage{
normalizeMatrix(exprs, study, metaData)
}
\arguments{
\item{exprs}{data.table of expression}

\item{study}{study accession eg \code{SDY269}}

\item{metaData}{list of study-specific meta data}
}
\description{
Normalize Matrix
}
