% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{orig_btm_list}
\alias{orig_btm_list}
\title{Original HIPC BTM list}
\format{
A list of lists with charactar vector elements
\describe{
\item{names}{geneSet pathway names}
\item{elements}{character vector of gene symbols}
...
}
}
\usage{
orig_btm_list
}
\description{
A dataset containing the geneSet pathways and geneIds for
a BTM file with the original date as October 08, 2013 generated
by the HIPC collaborators and stored in the HIPC google drive.
}
\keyword{datasets}
