% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{orig_emory}
\alias{orig_emory}
\title{Blood transcriptome modules}
\format{
A \code{list} of \code{character} of length 346, with one element per
module
}
\usage{
data(orig_emory)
}
\description{
Blood transcriptome modules
}
\references{
http://www.immuneprofiling.org/meni/meni-paper/btm-landing.gsp
http://www.nature.com/ni/journal/v15/n2/full/ni.2789.html
}
\keyword{datasets}
