% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_response_predictor.R
\name{print.ImmuneResponsePredictor}
\alias{print.ImmuneResponsePredictor}
\title{Print an ImmuneResponsePredictor object}
\usage{
\method{print}{ImmuneResponsePredictor}(x, ...)
}
\arguments{
\item{x}{an object of class ImmuneResponsePredictor}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print an ImmuneResponsePredictor object
}
