% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_files.R
\name{retrieve_input_files}
\alias{retrieve_input_files}
\title{Retrieve input files}
\usage{
retrieve_input_files(
  study,
  gef,
  meta_data,
  analysis_dir,
  verbose = FALSE,
  reload = FALSE
)
}
\arguments{
\item{study}{Study accession}

\item{gef}{gene expression files}

\item{meta_data}{List of study-specific metadata}

\item{analysis_dir}{directory to save intermediate files.}

\item{verbose}{print verbose logging statements?}

\item{reload}{re-download files if already found?}
}
\description{
Identify input file location and download and prepare
input files as needed, into a format that is standard per platform
}
