% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCreateMx.R
\name{runCreateMx}
\alias{runCreateMx}
\title{Run Create Matrix}
\usage{
runCreateMx(
  labkey.url.base,
  labkey.url.path,
  pipeline.root,
  analysis.directory,
  output.tsv,
  selectedBiosamples,
  fasId,
  taskOutputParams,
  onCL = FALSE
)
}
\arguments{
\item{labkey.url.base}{labkey.url.base}

\item{pipeline.root}{pipeline.root}

\item{analysis.directory}{analysis.directory}

\item{output.tsv}{Name of output matrix}

\item{selectedBiosamples}{List of biosamples (should be from one
cell type from one cohort)}

\item{fasId}{Feature Annotation Set ID}

\item{taskOutputParams}{taskOutputParams}

\item{onCL}{running on command line? (T/F)}

\item{labeky.url.path}{labkey.url.path}
}
\description{
Runs all steps of creating matrix in ImmuneSpace.
}
