% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_by_gene_symbol}
\alias{summarize_by_gene_symbol}
\title{Summarize Matrix}
\usage{
summarize_by_gene_symbol(
  ge_dt,
  feature_gene_map,
  method = "mean",
  verbose = FALSE
)
}
\arguments{
\item{ge_dt}{data.table of gene expression values. Features should be listed in
\code{feature_id} column, with additional columns for each sample. Samples
should be labelled by biosample id.}

\item{feature_gene_map}{data.table mapping original gene id (eg probe id, ensembl id, gene alias)
to gene symbol. It should have two columns: \code{featureid} (original) and \code{genesymbol} (updated).}

\item{method}{Summarization method. \code{mean} or \code{max}.
\code{mean} will return the mean of all features mapping
to a gene alias for each sample. This is the method used in the ImmuneSpace
HIPCMatrix pipeline.
\code{max} will chose the probe which has the greatest geometric mean expression
across all samples. This is the method used in the ImmuneSignatures analysis.}

\item{verbose}{write verbose logging statements?}
}
\description{
Summarize any duplicated genes so that there is
one entry per gene.
}
\details{
This function summarizes normalized expression values by mean
into a table with one entry per gene name based on \code{feature_gene_map}.
The input data should be in log2 space.
}
