% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCreateMx.R
\name{summarizeMatrix}
\alias{summarizeMatrix}
\title{Summarize Matrix}
\usage{
summarizeMatrix(norm_exprs, f2g)
}
\arguments{
\item{norm_exprs}{data.table of normalized gene expression values}

\item{f2g}{data.table mapping original gene id (eg probe id, ensembl id, gene alias)
to gene symbol. It should have two columns: \code{featureid} (original) and \code{genesymbol} (updated).}
}
\description{
Summarize any duplicated genes so that there is
one entry per gene.
}
\details{
This function summarizes normalized expression values into a table
with one entry per gene name by:
\enumerate{
\item Map original feature names to gene symbols based on \code{f2g} table
\item Remove any features that did not map to a gene symbol
\item Summarize duplicated gene symbols by mean of the (normalized) values.
}
}
