% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_response_predictor.R
\name{train_immune_response_predictors}
\alias{train_immune_response_predictors}
\title{Get Immune Response Predictors}
\usage{
train_immune_response_predictors(
  con,
  cohorts,
  timepoint,
  assay = "hai",
  timepoint_unit = "Days",
  use_only_de_genes = timepoint > 0,
  fc_thresh = 0.58,
  dichotomize = FALSE,
  dichotomize_thresh = 4,
  reload = FALSE
)
}
\description{
Get Immune Response Predictors
}
