% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\name{update_geneset_symbols}
\alias{update_geneset_symbols}
\title{Update Geneset Symbols}
\usage{
update_geneset_symbols(geneset_name, gene_alias_map = hgncAlias2Symbol)
}
\arguments{
\item{geneset_name}{name of a gene set found in HIPCMatrix package data}

\item{gene_alias_map}{mapping of gene symbol to alias.}
}
\description{
Update gene symbols in a geneset found in HIPCMatrix package data
}
