% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{updated_btm_df}
\alias{updated_btm_df}
\title{Updated HIPC BTM data frame}
\format{
A dataframe with pathways and geneIds
\describe{
\item{pathway}{geneSet pathway names}
\item{SYMBOL}{updated gene symbols}
...
}
}
\usage{
updated_btm_df
}
\description{
A dataset containing the geneSet pathways and geneIds for
a BTM file with the original date as October 08, 2013 generated
by the HIPC collaborators and stored in the HIPC google drive.
}
\keyword{datasets}
