% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_matrix}
\alias{write_matrix}
\title{Write matrix to file system}
\usage{
write_matrix(
  output_dir,
  matrix_name,
  exprs,
  norm_exprs,
  sum_exprs,
  verbose = FALSE
)
}
\arguments{
\item{output_dir}{path to output directory (directory where final matrices
will be written)}

\item{matrix_name}{Name of matrix}

\item{exprs}{data.table of raw (background-corrected) expression (from \code{makeRawMatrix()})}

\item{norm_exprs}{data.table of normalized expression (from \code{normalizeMatrix()})}

\item{sum_exprs}{data.table of summarized expression (from \code{summarizeMatrix()})}

\item{verbose}{Write verbose print statements?}
}
\description{
Write matrix to file system
}
\details{
Writes four versions of flat tsv files:
\enumerate{
\item \code{<matrix_name>.raw.tsv}: raw, background-corrected values
\item \code{<matrix_name>.tsv}: normalized values
\item \code{<matrix_name>.summary.tsv}: normalized values, summarized by gene symbol
(based on current annotation)
\item \code{<matrix_name>.summary.orig}: normalized values, summarized by gene symbol
(based on original annotation)
}
}
