% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCreateMx.R
\name{writeMatrix}
\alias{writeMatrix}
\title{Write matrix to file system}
\usage{
writeMatrix(
  pipeline.root,
  output.tsv,
  exprs,
  norm_exprs,
  sum_exprs,
  onCL = FALSE
)
}
\arguments{
\item{pipeline.root}{pipeline.root}

\item{output.tsv}{Name of matrix}

\item{exprs}{data.table of raw (background-corrected) expression (from \code{makeRawMatrix()})}

\item{norm_exprs}{data.table of normalized expression (from \code{normalizeMatrix()})}

\item{sum_exprs}{data.table of summarized expression (from \code{summarizeMatrix()})}

\item{onCL}{running on command line? (T/F)}
}
\description{
Write matrix to file system
}
\details{
Writes four verisons of flat tsv files:
\enumerate{
\item \code{<matrix_name>.raw.tsv}: raw, background-corrected values
\item \code{<matrix_name>.tsv}: normalized values
\item \code{<matrix_name>.summary.tsv}: normalized values, summarized by gene symbol
(based on current annotation)
\item \code{<matrix_name>.summary.orig}: normalized values, summarized by gene symbol
(based on original annotation)
}
}
