% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOUtils.R
\name{.make_id_to_gsm_map}
\alias{.make_id_to_gsm_map}
\title{Make id to GSM map}
\usage{
.make_id_to_gsm_map(
  gsms,
  study_id_term,
  gsm_map_index = 1,
  id_regex_map = NULL,
  supp_files_dir = NULL
)
}
\arguments{
\item{gsms}{Input vector of gsm accession}

\item{study_id_term}{Term from getGEO(gsm) header to use for extracting
sample id. "description" or "title"}

\item{gsm_map_index}{Index returning id in GSM header}

\item{id_regex_map}{list specifying regex terms to map GSE id's using gsub
of the form \code{list(old = "", new = "")}}

\item{supp_files_dir}{path to directory to save supplementary files.}
}
\description{
Create map of GSM accessions to ID's provided in GSE. If needed, appropriate
info for the matrix will be returned by \code{get_meta_data}.
}
