% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raw_matrix.R
\name{.map_feature_id_col}
\alias{.map_feature_id_col}
\title{Standardize probe column name}
\usage{
.map_feature_id_col(exprs_dt)
}
\arguments{
\item{exprs_dt}{data.table of gene expression with one column per sample,
one row per feature.}
}
\description{
Different platforms use different names for the feature id column.
This will update this column to be called \code{feature_id}
}
