% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOUtils.R
\name{.prep_geo_files}
\alias{.prep_geo_files}
\title{Prep GEO files}
\usage{
.prep_geo_files(study, gef, metaData, input_files)
}
\arguments{
\item{study}{study accession eg \code{SDY269}}

\item{gef}{result of ISCon$getDataset("gene_expression_files") for one run.}

\item{metaData}{list of study-specific meta data}

\item{input_files}{input file names}
}
\value{
path to raw, prepped input files. tsv for everyone except affy, which
will be the path to CEL files.
}
\description{
Generate flat files that are ready for processing from GEO "raw" data.
Warning! Raw data is highly variable for gse supplementary files
}
