% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmPortUtils.R
\name{.prep_immport_files}
\alias{.prep_immport_files}
\title{Prep ImmPort Files}
\usage{
.prep_immport_files(study, gef, platform, input_files, analysis_dir)
}
\arguments{
\item{study}{study accession eg \code{SDY269}}

\item{gef}{result of ISCon$getDataset("gene_expression_files") for one run.}

\item{input_files}{input file names}

\item{meta_data}{list of study-specific meta data}
}
\value{
path to raw, prepped input files
}
\description{
Prep ImmPort Files
}
