% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmx_differential_expression.R
\name{find_de_genes_eBayes}
\alias{find_de_genes_eBayes}
\title{Find Differentially Expressed genes using Empirical Bayes Statistics}
\usage{
find_de_genes_eBayes(eset, p_val_cutoff = 0.02)
}
\arguments{
\item{eset}{expressionset object with normalized microarray expression.}

\item{p_val_cutoff}{adjusted p-value cutoff for determining differentially
expressed genes.}
}
\value{
list with one entry for each non-baseline timepoint
}
\description{
Compute differential expression over time on microarray data using
empirical Bayes statistics on a linear model using the
\code{\link[limma]{eBayes}} method from the \code{limma} package. All
differentially expressed genes for each timepoint are reported.
}
