% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_meta_data}
\alias{get_meta_data}
\title{Get Meta Data}
\usage{
get_meta_data(
  study,
  gef = NULL,
  fas_id = NULL,
  baseUrl = "https://www.immunespace.org"
)
}
\arguments{
\item{study}{study accession}

\item{gef}{gene expression files. Only required for SDY1529.}

\item{fas_id}{Feature Annotation Set ID (from FeatureAnnotationSet table).
If not provided, will attempt to derive based on study. Must be provided
if study has not already been processed.}

\item{baseUrl}{Base URL to pull feature annotation info from.}
}
\description{
Get Meta Data
}
