% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_meta_data}
\alias{get_meta_data}
\title{Get Meta Data}
\usage{
get_meta_data(
  study,
  gef = NULL,
  fas_id = NULL,
  baseUrl = "https://www.immunespace.org"
)
}
\arguments{
\item{study}{study accession}

\item{gef}{gene expression files. Only required for SDY1529.}

\item{fas_id}{Feature Annotation Set ID (from FeatureAnnotationSet table).
If not provided, will attempt to derive based on study. Must be provided
if study has not already been processed.}

\item{baseUrl}{Base URL to pull feature annotation info from.}
}
\value{
List of meta data detailing how to process matrix. Includes:
\describe{
\item{file_location}{Source for raw files. Options are \code{gsm_soft}:
in data returned with \code{GEOquery::getGEO()}, \code{gsm_supp_files}:
supplementary files in sample entry (GSM), \code{gse_supp_files}:
supplementary files in experement entry (GSE), \code{immport}: raw files
provided from ImmPort, \code{custom}: somewhere else (generally provided
by immport as they wait for an update to make its way to GEO/ImmPort)}
\item{platform}{Sequencing platform: Either Affymetrix, Illumina, Stanford
(custom platform), or NA (RNA-seq). TODO: Add more descriptive identifier
for RNA-seq}
\item{id_to_gse_mapping_info (optional)}{Mapping info for data found in GSE
to map ids to GSM (see \href{https://github.com/RGLab/HIPCMatrix/blob/main/data-raw/meta_data_list.R}{source}
for more details)}
\item{custom_file_info (optional)}{Details on where to find custom file for
when \code{file_location == "custom"}}
\item{illumina_manifest_file (optional)}{Path to manifest file for studies with Illumina
idat files that need bgx manifest files.}
\item{gsm_table_var_name (optional)}{Custom list of raw values column name for gsm-based data}
}
}
\description{
Get meta data about how to process matrices within a study or
how a matrix was processed.
}
\details{
Details such as \code{file_location} must be added manually
to the \code{\link{meta_data_list}} object for new matrices. See
\href{https://www.notion.so/rglab/HIPCMatrix-meta_data_list-0e501a51551843b78f68806f1d705ec1}{internal Notion documentation}
for more details on how to update.
}
