% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmx_feature_annotation.R
\name{mapAlias2Symbol}
\alias{mapAlias2Symbol}
\title{map alias to gene symbol}
\usage{
mapAlias2Symbol(aliases, gene_symbol_map = hgncAlias2Symbol)
}
\arguments{
\item{aliases}{character vector of gene aliases}

\item{gene_symbol_map}{data.table with gene aliases and matching symbol.
Columns should be "ALIAS" and "SYMBOL". Default is to use hgncAlias2Symbol
object provided in package data.}
}
\description{
Function to update annotation of a character vector of gene aliases, using
the mapping provided by \code{\link{hgncAlias2Symbol}}
}
