% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{normalize_matrix}
\alias{normalize_matrix}
\title{Normalize matrix}
\usage{
normalize_matrix(exprs_dt, metaData)
}
\arguments{
\item{exprs_dt}{data.table with raw, background-corrected expression. One
column per sample, one row per feature.}

\item{metaData}{list of study-specific meta data}
}
\description{
Normalize matrix
}
