% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{normalize_microarray}
\alias{normalize_microarray}
\title{Normalize microarray data}
\usage{
normalize_microarray(
  exprs_mx,
  log2_transform = TRUE,
  force = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{exprs_mx}{matrix with background-corrected microarray
expression data, one column per sample, one row per feature.}

\item{log2_transform}{(boolean) Should log2 transform be performed before normalization?
This should be FALSE for affy data which was read using RMA, which is already
in log2 scale.}

\item{force}{(boolean) Force log2 transform. If \code{log2_transform} is
\code{TRUE}, this function will error if \code{counts_dt} does not have
values greater than 100, unless \code{force = TRUE}.}
}
\value{
a matrix with one row per feature and one column per sample.
}
\description{
Normalize across samples to correct for technical variation using
quantile normalization. Also log2-transforms the data.
}
