% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_rnaseq}
\alias{normalize_rnaseq}
\title{Normalize raw RNA-seq data}
\usage{
normalize_rnaseq(counts_mx, verbose = FALSE)
}
\arguments{
\item{counts_mx}{matrix with raw counts, one column per sample, one
row per feature}
}
\value{
a matrix with one row per feature and one column per sample.
}
\description{
RNA data will be normalized using the variance stabilizing transformation,
as implemented by the \code{DESeq2::vst} function.
}
