% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\docType{data}
\name{updated_btm_list}
\alias{updated_btm_list}
\title{Updated HIPC BTM list}
\format{
A list of lists with character vector elements
\describe{
\item{names}{geneSet pathway names}
\item{elements}{character vector of gene symbols}
...
}
}
\usage{
updated_btm_list
}
\description{
A dataset containing the geneSet pathways and geneIds for
a BTM file that is updated using the most recent version of
the HGNC database when package data is re-built.
}
\keyword{datasets}
